+++ songsServlet/pom.xml
@@ -24,13 +24,6 @@
             <version>4.12</version>
         </dependency>
 
-        <!-- https://mvnrepository.com/artifact/commons-cli/commons-cli -->
-        <dependency>
-            <groupId>commons-cli</groupId>
-            <artifactId>commons-cli</artifactId>
-            <version>1.4</version>
-        </dependency>
-
         <dependency>
             <groupId>org.hamcrest</groupId>
             <artifactId>hamcrest-library</artifactId>
@@ -98,7 +91,7 @@
                     <!-- MainClass in mainfest make a executable jar -->
                     <archive>
                         <manifest>
-                            <mainClass>de.htw.ai.kbe.servlet.MyServlet</mainClass>
+                            <mainClass>de.htw.ai.kbe.servlet.Servlet</mainClass>
                         </manifest>
                     </archive>
