+++ songsServlet/src/main/java/de/htw/ai/kbe/servlet/Servlet.java
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class MyServlet extends HttpServlet {
+public class Servlet extends HttpServlet {
     //    public static void main(String[] args) {
 //        System.out.println("coming soon");
 //    }
